move stack template to stack_template.go
